requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyYAML==5.1.2
toml==0.10.0
Werkzeug==0.16.0
